Initialization Bug Check
On startup, workshop time is not initialized and for some reason, this means the map time and all other form of timers will restart to 0 once that time is set to a value. The extra check prevents the new restart functionality from working until the workshop time is initialized (aka a level is finished). This is a tiny bit of a bug with the autosplitter, but to be fair it's only a bug with the new restart functionality itself and now shouldn't affect anyone in a meaningful way. 

I don't understand why Clustertruck has weird value initialization on the first run, but whatever.