Category Extension Settings
Reset in menu is a specific setting (only works if reset is enabeld). Also have settings for AllLevels category so that split by world works with holidays (by counting the first 5 splits and splitting when it hits 5). Also did a fix to unpause the autosplitter when in the menu (it shouldn't be paused when in the menu).