General Refactoring and Deathless/All Levels
All levels functionality is still glitched and I am not sure why. I don't have time to work on this anymore and can't get it to work right now. What this needs is a state machine, not cleaner if statements. You can clean a pile of shit, but its still shit.